FCFIELDS-34: ModuleName overwrites actual ModuleName
folio-custom-fields ships with ModuleName.java that contains the module name "folio_custom_fields".

mod-users has it's own ModuleName.java with module name mod_users.

Which one is used depends on the order the jar files are returned by the file system.

When running in my IDE sometimes mod-users has the wrong one.

folio-custom-fields is a library that doesn't have a module name and therefore it's a bug that it ships with ModuleName.