FCFIELDS-68: Run removeValues in transaction
https://folio-org.atlassian.net/browse/FCFIELDS-68
https://folio-org.atlassian.net/browse/FCFIELDS-61

Purpose

The problem is that removeValues (which delegates to RecordService) does not support passing a connection object, and therefore executes its database operations using a different connection from the one used in the transaction block.

This results in database locks.

Approach

Run removeValues in the transaction.

TODOs and Open Questions
- [x] Check logging.